refactor(test): move integration test for `mysql-async` to a separate crate
