+++ ci/scripts/e2e-source-test.sh
@@ -80,8 +80,8 @@ echo "--- e2e, ci-1cn-1fe, mysql & postgres cdc"
 # import data to mysql
 mysql --host=mysql --port=3306 -u root -p123456 < ./e2e_test/source_legacy/cdc/mysql_cdc.sql
 
-echo "run single mysql unit test"
-cargo test --package risingwave_connector --lib -- --ignored test_mysql_async_with_connection_pool
+echo "run mysql-async integration test"
+cargo test --package risingwave_mysql_test -- --ignored
 # import data to postgres
 export PGHOST=db PGPORT=5432 PGUSER=postgres PGPASSWORD='post\tgres' PGDATABASE=cdc_test
 createdb
