+++ src/tests/mysql_test/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "risingwave_mysql_test"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+
+[dev-dependencies]
+tokio = { version = "0.2", package = "madsim-tokio", features = [
+    "fs",
+    "rt",
+    "rt-multi-thread",
+    "sync",
+    "macros",
+    "time",
+    "signal",
+] }
+mysql_async = { workspace = true }
+futures = "0.3"
+
+[lints]
+workspace = true
